requirements.txt
python-telegram-bot==13.15
requests==2.32.3
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
pyarrow==15.0.2
